Implement more counters modes.
New counters were implemented:
  - bytes - calculate amount of bytes allocated by each function.
  - ticks - calculate number of ticks spent in each function.

Main rationale is that alloc and time in percentage is fairily
unusable in non-trivial programs or programs that spends much time
in idle. So both bytes and ticks gives better stats.

Additional modes require program to be run with -P, so additional
checks were introduced in order to notify user about this requirement.